requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
openai==1.54.0
httpx<0.28.0  # Required for compatibility with openai 1.54.0
gunicorn==21.2.0  # WSGI server for production deployment
psycopg2-binary==2.9.9  # PostgreSQL adapter for Azure Cosmos DB for PostgreSQL

# Background task processing
celery==5.3.4
redis==5.0.1
flower==2.0.1  # Celery monitoring dashboard
